requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlmodel>=0.0.8
asyncpg>=0.28.0
python-multipart>=0.0.6
pydantic-settings>=2.0.0
google-auth>=2.22.0
requests>=2.31.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Test Dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.24.1
aiosqlite>=0.17.0
pytest-env>=0.8.0
